feat: 添加热门帖子功能

+++ src/main/java/com/example/entity/dao/TopicDO.java
@@ -23,6 +23,6 @@ public class TopicDO{
     private Integer type;
     private Integer top;
     private Date time;
-    private Double score;
+    private Long score;
     private Integer status; // 0 表示草稿，1 表示已发布
 }
